# backend/test/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/speakers": {
      "post": {
        "operationId": "SpeakersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpeakerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Speakers"
        ]
      },
      "get": {
        "operationId": "SpeakersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by. Each endpoint documents its own allowed values.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Speakers"
        ]
      }
    },
    "/speakers/{id}": {
      "get": {
        "operationId": "SpeakersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Speakers"
        ]
      },
      "patch": {
        "operationId": "SpeakersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpeakerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Speakers"
        ]
      }
    },
...
    "/events/{eventId}/incidents": {
      "post": {
        "operationId": "IncidentsController_create",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIncidentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Incidents"
        ]
      }
    },
    "/incidents": {
      "get": {
        "operationId": "IncidentsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by. Each endpoint documents its own allowed values.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "event",
            "required": false,
            "in": "query",
            "description": "Filter by event id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignee",
            "required": false,
            "in": "query",
            "description": "Filter by assignee user id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "severity",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "critical",
                "high",
                "medium",
                "low"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "investigating",
                "mitigated",
                "resolved"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Incidents"
        ]
      }
    },
    "/incidents/{id}": {
      "get": {
        "operationId": "IncidentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Incidents"
        ]
      },
      "patch": {
        "operationId": "IncidentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIncidentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Incidents"
        ]
      }
    },
    "/incidents/{id}/status": {
      "patch": {
        "operationId": "IncidentsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIncidentStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Incidents"
        ]
      }
    },
    "/incidents/{id}/timeline": {
      "get": {
        "operationId": "IncidentsController_getTimeline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Incidents"
        ]
      }
    },
...
  },
  "info": {
    "title": "EventOps Intelligence Platform API",
    "description": "Internal event-operations platform API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSpeakerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ada Lovelace"
          },
          "title": {
            "type": "string",
            "example": "Principal Engineer, Acme Corp"
          },
          "bio": {
            "type": "string",
            "example": "Ada has 15 years of experience in..."
          },
          "email": {
            "type": "string",
            "example": "ada@example.com"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateSpeakerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ada Lovelace"
          },
          "title": {
            "type": "string",
            "example": "Principal Engineer, Acme Corp"
          },
          "bio": {
            "type": "string",
            "example": "Ada has 15 years of experience in..."
          },
          "email": {
            "type": "string",
            "example": "ada@example.com"
          }
        }
      },
...
      "CreateIncidentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Main stage sound system failure"
          },
          "description": {
            "type": "string",
            "example": "Audio cut out mid-keynote, backup mics not working"
          },
          "session": {
            "type": "string",
            "description": "Optional session this incident occurred during"
          },
          "severity": {
            "type": "string",
            "enum": [
              "critical",
              "high",
              "medium",
              "low"
            ]
          },
          "assignee": {
            "type": "string",
            "description": "Operations Member user id"
          }
        },
        "required": [
          "title",
          "severity"
        ]
      },
      "UpdateIncidentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Main stage sound system failure"
          },
          "description": {
            "type": "string",
            "example": "Audio cut out mid-keynote, backup mics not working"
          },
          "session": {
            "type": "string",
            "description": "Optional session this incident occurred during"
          },
          "severity": {
            "type": "string",
            "enum": [
              "critical",
              "high",
              "medium",
              "low"
            ]
          },
          "assignee": {
            "type": "string",
            "description": "Operations Member user id"
          }
        }
      },
      "UpdateIncidentStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "investigating",
              "mitigated",
              "resolved"
            ]
          }
        },
        "required": [
          "status"
        ]
      }
    }
  }
}